+++ daskms/reads.py
@@ -177,7 +177,7 @@ def getter_wrapper(row_orders, *args):
 
 
 def _dataset_variable_factory(table_proxy, table_schema, select_cols,
-                              exemplar_row, orders, chunks, array_prefix):
+                              exemplar_row, orders, chunks, array_suffix):
     """
     Returns a dictionary of dask arrays representing
     a series of getcols on the appropriate table.
@@ -200,7 +200,7 @@ def _dataset_variable_factory(table_proxy, table_schema, select_cols,
         appropriate rows to extract from the table.
     chunks : dict
         Chunking strategy for the dataset.
-    array_prefix : str
+    array_suffix : str
         dask array string prefix
 
     Returns
@@ -234,6 +234,11 @@ def _dataset_variable_factory(table_proxy, table_schema, select_cols,
                 args.append(dim_extents_array(d, c))
                 args.append((d,))
 
+            # Disable getcolslice caching
+            # https://github.com/ska-sa/dask-ms/issues/92
+            # https://github.com/casacore/casacore/issues/1018
+            table_proxy.setmaxcachesize(column, 1).result()
+
             new_axes = {}
         else:
             # We need to inform blockwise about the size of our
@@ -248,7 +253,7 @@ def _dataset_variable_factory(table_proxy, table_schema, select_cols,
 
         # Name of the dask array representing this column
         token = dask.base.tokenize(args)
-        name = "-".join((array_prefix, column, token))
+        name = "~".join(("read", column, array_suffix)) + "-" + token
 
         # Construct the array
         dask_array = da.blockwise(getter_wrapper, full_dims,
@@ -293,11 +298,12 @@ class DatasetFactory(object):
         self.taql_where = kwargs.pop('taql_where', '')
         self.table_keywords = kwargs.pop('table_keywords', False)
         self.column_keywords = kwargs.pop('column_keywords', False)
+        self.table_proxy = kwargs.pop('table_proxy', False)
 
         if len(kwargs) > 0:
             raise ValueError("Unhandled kwargs: %s" % kwargs)
 
-    def _table_proxy(self):
+    def _table_proxy_factory(self):
         return TableProxy(pt.table, self.table_path, ack=False,
                           readonly=True, lockoptions='user',
                           __executor_key__=executor_key(self.canonical_name))
@@ -305,11 +311,10 @@ class DatasetFactory(object):
     def _table_schema(self):
         return lookup_table_schema(self.canonical_name, self.table_schema)
 
-    def _single_dataset(self, orders, exemplar_row=0):
+    def _single_dataset(self, table_proxy, orders, exemplar_row=0):
         _, t, s = table_path_split(self.canonical_name)
         short_table_name = "/".join((t, s)) if s else t
 
-        table_proxy = self._table_proxy()
         table_schema = self._table_schema()
         select_cols = set(self.select_cols or table_proxy.colnames().result())
         variables = _dataset_variable_factory(table_proxy, table_schema,
@@ -326,10 +331,9 @@ class DatasetFactory(object):
 
         return Dataset(variables, coords=coords)
 
-    def _group_datasets(self, groups, exemplar_rows, orders):
+    def _group_datasets(self, table_proxy, groups, exemplar_rows, orders):
         _, t, s = table_path_split(self.canonical_name)
         short_table_name = '/'.join((t, s)) if s else t
-        table_proxy = self._table_proxy()
         table_schema = self._table_schema()
 
         datasets = []
@@ -353,7 +357,7 @@ class DatasetFactory(object):
 
             # Prefix d
             gid_str = ",".join(str(gid) for gid in group_id)
-            array_prefix = "%s-[%s]" % (short_table_name, gid_str)
+            array_suffix = "[%s]-%s" % (gid_str, short_table_name)
 
             # Create dataset variables
             group_var_dims = _dataset_variable_factory(table_proxy,
@@ -361,7 +365,7 @@ class DatasetFactory(object):
                                                        select_cols,
                                                        exemplar_row,
                                                        order, group_chunks,
-                                                       array_prefix)
+                                                       array_suffix)
 
             # Extract ROWID
             try:
@@ -381,14 +385,14 @@ class DatasetFactory(object):
         return datasets
 
     def datasets(self):
-        table_proxy = self._table_proxy()
+        table_proxy = self._table_proxy_factory()
 
         # No grouping case
         if len(self.group_cols) == 0:
             order_taql = ordering_taql(table_proxy, self.index_cols,
                                        self.taql_where)
             orders = row_ordering(order_taql, self.index_cols, self.chunks[0])
-            datasets = [self._single_dataset(orders)]
+            datasets = [self._single_dataset(table_proxy, orders)]
         # Group by row
         elif len(self.group_cols) == 1 and self.group_cols[0] == "__row__":
             order_taql = ordering_taql(table_proxy, self.index_cols,
@@ -408,7 +412,8 @@ class DatasetFactory(object):
             # dataset as an attribute
             np_sorted_row = sorted_rows.compute()
 
-            datasets = [self._single_dataset((row_blocks[r], run_blocks[r]),
+            datasets = [self._single_dataset(table_proxy,
+                                             (row_blocks[r], run_blocks[r]),
                                              exemplar_row=er)
                         for r, er in enumerate(np_sorted_row)]
         # Grouping column case
@@ -422,7 +427,8 @@ class DatasetFactory(object):
             exemplar_rows = order_taql.getcol("__firstrow__").result()
             assert len(orders) == len(exemplar_rows)
 
-            datasets = self._group_datasets(groups, exemplar_rows, orders)
+            datasets = self._group_datasets(table_proxy, groups,
+                                            exemplar_rows, orders)
 
         ret = (datasets,)
 
@@ -432,6 +438,9 @@ class DatasetFactory(object):
         if self.column_keywords is True:
             keywords = table_proxy.submit(_col_keyword_getter, READLOCK)
             ret += (keywords.result(),)
+
+        if self.table_proxy is True:
+            ret += (table_proxy,)
 
         if len(ret) == 1:
             return ret[0]