+++ daskms/tests/test_dataset_keywords.py
@@ -5,7 +5,13 @@ import pyrap.tables as pt
 import pytest
 
 from daskms.example_data import example_ms
+from daskms.table_proxy import TableProxy
 from daskms import xds_to_table, xds_from_ms
+
+try:
+    from xarray import Dataset
+except ImportError:
+    from daskms.dataset import Dataset
 
 
 @pytest.fixture(scope='module')
@@ -18,14 +24,16 @@ def keyword_ms():
 
 @pytest.mark.parametrize("table_kw", [True, False])
 @pytest.mark.parametrize("column_kw", [True, False])
-def test_keyword_read(keyword_ms, table_kw, column_kw):
+@pytest.mark.parametrize("table_proxy", [True, False])
+def test_read_keywords(keyword_ms, table_kw, column_kw, table_proxy):
     # Create an example MS
     with pt.table(keyword_ms, ack=False, readonly=True) as T:
         desc = T._getdesc(actual=True)
 
     ret = xds_from_ms(keyword_ms,
                       table_keywords=table_kw,
-                      column_keywords=column_kw)
+                      column_keywords=column_kw,
+                      table_proxy=table_proxy)
 
     if isinstance(ret, tuple):
         ret_pos = 1
@@ -41,17 +49,26 @@ def test_keyword_read(keyword_ms, table_kw, column_kw):
                 assert desc[column]['keywords'] == keywords
 
             ret_pos += 1
+
+        if table_proxy is True:
+            tp = ret[ret_pos]
+            assert isinstance(tp, TableProxy)
+            assert tp.nrows().result() == 10
+
+            ret_pos += 1
     else:
         assert table_kw is False
         assert column_kw is False
+        assert table_proxy is False
         assert isinstance(ret, list)
 
 
-def test_keyword_write(ms):
+def test_write_keywords(ms):
     datasets = xds_from_ms(ms)
 
     # Add to table keywords
     writes = xds_to_table([], ms, [], table_keywords={'bob': 'qux'})
+    assert isinstance(writes, Dataset)
     dask.compute(writes)
 
     with pt.table(ms, ack=False, readonly=True) as T:
@@ -60,6 +77,7 @@ def test_keyword_write(ms):
     # Add to column keywords
     writes = xds_to_table(datasets, ms, [],
                           column_keywords={'STATE_ID': {'bob': 'qux'}})
+    assert isinstance(writes, list) and isinstance(writes[0], Dataset)
     dask.compute(writes)
 
     with pt.table(ms, ack=False, readonly=True) as T:
@@ -70,8 +88,22 @@ def test_keyword_write(ms):
     writes = xds_to_table(datasets, ms, [],
                           table_keywords={'bob': DELKW},
                           column_keywords={'STATE_ID': {'bob': DELKW}})
+    assert isinstance(writes, list) and isinstance(writes[0], Dataset)
     dask.compute(writes)
 
     with pt.table(ms, ack=False, readonly=True) as T:
         assert 'bob' not in T.getkeywords()
         assert 'bob' not in T.getcolkeywords("STATE_ID")
+
+
+def test_write_table_proxy_keyword(ms):
+    datasets = xds_from_ms(ms)
+
+    # Test that we get a TableProxy if requested
+    writes, tp = xds_to_table(datasets, ms, [], table_proxy=True)
+    assert isinstance(writes, list) and isinstance(writes[0], Dataset)
+    assert isinstance(tp, TableProxy)
+    assert tp.nrows().result() == 10
+
+    writes = xds_to_table(datasets, ms, [], table_proxy=False)
+    assert isinstance(writes, list) and isinstance(writes[0], Dataset)